refactor(extension): extract connection and config helpers
